pyproject.toml
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"


[project]
name = "docx_comments"
version = "0.1.0"
authors = [
    { name = "Frank Male", email = "frank.male@psu.edu" },
]
maintainers = [
    { name = "Frank Male", email = "frank.male@psu.edu" },
]
license = { file = "LICENSE" }

description = "Gather comments from Word docx files"
readme = "README.md"

requires-python = ">=3.8"

classifiers = [
    "License :: OSI Approved :: BSD License",
    "Topic :: Scientific/Engineering",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Development Status :: 1 - Planning",
]

dependencies = [
    "lxml >=4.9.1",
    "click >=8.1",
]

[project.optional-dependencies]
test = [
    "pytest >=6",
    "python-dateutil >=2.8"
]
dev = [
    "pytest >=6",
]
docs = [
    "Sphinx>=4.0",
     "myst_parser>=0.13",
    "sphinx-book-theme>=0.1.0",
    "sphinx_copybutton",
]

[project.urls]
homepage = "https://github.com/frank10101111/docx-comments"
[tool.pdm.dev-dependencies]
devtest = ["pytest"]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = ["error"]
testpaths = [
    "tests",
]


[tool.mypy]
files = "src"
python_version = "3.8"
warn_unused_configs = true
strict = true
show_error_codes = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
ignore_missing_imports = true


[tool.isort]
profile = "black"


[tool.pylint]
master.py-version = "3.7"
master.ignore-paths= ["src/docx_comments/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "wrong-import-position",
]
